Refactor AuthHandler code + restore getJWT method
After some early snafus in the code to get JWTs, we had restored hardcoded
authorization in which we obtained the authorization as part of the remote
call.

However, the snafu was caused by an actual bug in the code - when the token was
NULL, we would read it from the keychain, but then return with an error
immediately instead of attempting to refresh it.

Part of the reason for this snafu was the horribly tangled code, which was
written at a time when I just wanted to get it to work so that we could get our
app into the store, and I didn't really understand the flow very well.

Restructured the entire code into something much easier to understand, split
between UI-based sign in and automatic background authentication, and fixed the
bug as well.

Also, changed the code to integrate with the most recent version of the google
OAuth library, that uses the new GTMSessionFetcher instead of the old
GTMHTTPFetcher. Without this fix, background refresh of the tokens would hang
until the next push.