Completely restructure and simplify the AuthCompletionHandler
Since we are using the google sign in library, we no longer need to track and
refresh the tokens manually. Instead, the library runs the refresh for us. But,
since we want to know when the refresh is complete so that we can push changes
to the server, we need to convert the delegate based mechanism into a
callback-block based mechanism.

This change basically does that. It completely changes the
`AuthCompletionHandler` to really just handle the completion. And it changes
the plugin to use the new handler.

It also generates a notification if you are not logged in, which will help with
the transition from the old sign-in method to the new one.

Testing for the transition:

```
Installed version in app store
Waited until the end of the hour
Confirmed that we got a successful notification and a successful push
Ran the new version
"Error Loading Data"
Got new notification, cleared it
Sent to background (unclear if upgrade process from store launches app or not,
    suspect not)
Haven't re-logged in
Leave it running overnight
Update failed, unsure whether token has changed and whether we will get notifications
No notifications - debug build
Went to school - tracking started and stopped, but no notification because app
was in background, not launched
Force killed app
Came home
Got notification, logged in
Data pushed successfully
~ 2 hours later went to hardware store
Data pushed successfully twice
```