Some more cleanup of the creator interface
The main difference is that instead of passing the activity (on android) or the
viewcontroller (on iOS), we pass the plugin as an argument to the token
creator, *for the signIn case*.

For anything other than the sign in case, we don't actually need any UI stuff,
so this allows us to avoid the awkward instanceof check for activity on
android, and the equally awkward need to remember to set the view controller on iOS.

And passing in the plugin also allows us to set activity result callback from
the google-specific implementation instead of the generic plugin (we may not
use the activity callbacks for all implementations) and makes it easier to
implement the dev implementation that needs to execute javascript.

Includes minor fix to the way that files are copied over.