Add support for a simple dev mode authentication
In this, the user simply enters their email address. If the email address
matches one that was used to load data before this, the account is re-used,
which makes it easier to review and re-test data.

It also ensure that sessions are maintained across app restarts without logging
in every time, which makes debugging a lot easier.

And finally, it is an example of a javascript-based auth callback, which should
make it much easier to integrate with other auth systems if they don't already
have nice libraries to do it for us.

NOTE: the javascript stuff is extremely kludgy and ugly-looking, and I will
clean it up shortly after some more experimentation.