Change the auth check to only run after the intro is done
The javascript interface that sets the connection settings will run only after
all the plugins were loaded. But in plugin initialize, we were checking to see
if the login was failing even after the intro was done. So the first time that
the app was launched, the login code would run without any configuration.

Swap the order of the checks to avoid this issue.