Support specifying the prompt string for a generic `prompted auth`
The prompt string for a simple one-value "token" or "email" can now be
configured.  So we can use it for both "dummy-dev" and "token-list" modes, AND
the text can be in various languages if needed.

The files were renamed in 134aa5be3feb53d1d228c376080343421108709d
This is the actual changes to read the prompt and pass it in to the javascript,
which now takes the prompt as an argument.