Add an option to save programatically save local tokens
Sometimes, we don't want to prompt the user for a token but generate one in the
javascript and store it. The generating is easy enough, but the storage is
currently hard since it involves sending a custom URL notification to the plugin.

Instead of fiddling around with weird inapp browser schemes to pass through the URL,
let us handle this directly in the plugin since we wrote it ourselves.

+ bump up version number