Upgrade the version of GoogleSignIn to remove references to UIWebView
This also required an upgrade to AppAuth, and changes to the GoogleAuth and
OpenIDAuth code. This compiles, and will let me publish to the AppStore without
any UIWebView references, so I am going to check it in. However, I have not yet
tested the GoogleSigninAuth code, so this is tagged as an alpha release.