Modify prompted auth storage part 1 (get + migrate)
It turns out that the storage locations for iOS (a lot) and android (a little)
are not that reliable. We have had multiple iOS users and at least one android
user have their stored token deleted.

One of the iOS users said that she had problems with her phone so she had reset
to defaults. Maybe that is the problem?

We have seen this for only one android user so far, but it appears to happen on
android as well. Storing this in the database instead since that appears to be
more reliable.

This is the first part of the change which modifies the read method + handles a
transparent migration from the old storage location.

The migration essentially reads from the old location if the new location
doesn't exist and stores in the new location.

Testing done:

Ran this part alone on both android and iOS
Saw messages like:

```
2020-12-05 03:15:40.017419-0800 CanBikeCO[69221:59991767] Auth not found in local storage, copying from user profile
2020-12-05 03:15:40.017511-0800 CanBikeCO[69221:59991767] DEBUG: Auth not found in local storage, copying from user profile
2020-12-05 03:15:40.395587-0800 CanBikeCO[69221:59991767] Auth found in local storage, now it should be stable
2020-12-05 03:15:40.395679-0800 CanBikeCO[69221:59991767] DEBUG: Auth found in local storage, now it should be stable
```

and

```
2020-12-05 03:06:23.522 19156-19269/edu.berkeley.eecs.emission.devapp I/PromptedAuth: Auth not found in local storage, copying from user profile
2020-12-05 03:10:54.127 19518-19669/edu.berkeley.eecs.emission.devapp I/PromptedAuth: Auth found in local storage, now it should be stable
2020-12-05 03:10:54.373 19518-19669/edu.berkeley.eecs.emission.devapp I/PromptedAuth: Auth found in local storage, now it should be stable
```