Modify prompted auth storage part 2 (set)
It turns out that the storage locations for iOS (a lot) and android (a little)
are not that reliable. We have had multiple iOS users and at least one android
user have their stored token deleted.

One of the iOS users said that she had problems with her phone so she had reset
to defaults. Maybe that is the problem?

We have seen this for only one android user so far, but it appears to happen on
android as well. Storing this in the database instead since that appears to be
more reliable.

This is the second part of the change which modifies the write method. The iOS
implementation is trivial; the android implementation is a bit more complicated
since we need to add a try/catch block, and ensure that the
`setPromptedAuthToken` plugin method doesn't write the token directly.

Testing done:

Ran this part on both android and iOS
Saw messages like:

```
2020-12-05 04:32:30.667383-0800 CanBikeCO[72377:60060640] in handleNotification, feature = auth
2020-12-05 04:32:30.667467-0800 CanBikeCO[72377:60060640] DEBUG: in handleNotification, feature = auth
2020-12-05 04:32:30.678611-0800 CanBikeCO[72377:60060640] GOT URL:emission://auth?method=prompted-auth&token=djZsaYNh
2020-12-05 04:32:30.682104-0800 CanBikeCO[72377:60060640] Auth found in local storage, now it should be stable
2020-12-05 04:32:30.682192-0800 CanBikeCO[72377:60060640] DEBUG: Auth found in local storage, now it should be stable
2020-12-05 04:32:31.051128-0800 CanBikeCO[72377:60060640] Auth found in local storage, now it should be stable
```

```
2020-12-05 04:45:51.724 24197-24293/edu.berkeley.eecs.emission.devapp I/PromptedAuth: Auth found in local storage, now it should be stable
2020-12-05 04:45:51.808 24197-24293/edu.berkeley.eecs.emission.devapp I/PromptedAuth: Auth found in local storage, now it should be stable
2020-12-05 04:45:51.976 24197-24293/edu.berkeley.eecs.emission.devapp I/PromptedAuth: Auth found in local storage, now it should be stable
2020-12-05 04:45:52.126 24197-24391/edu.berkeley.eecs.emission.devapp I/PromptedAuth: Auth found in local storage, now it should be stable
2020-12-05 04:45:52.847 24197-24293/edu.berkeley.eecs.emission.devapp I/PromptedAuth: Auth found in local storage, now it should be stable
```