Fix callback handling through simplification
The change in
https://github.com/google/GoogleSignIn-iOS/blob/main/CHANGELOG.md#600-2021-7-13
indicates that the
the delegate has been replaced by callbacks.

This makes it much closer to the code pattern that we already have.

So we can remove the `registerCallback` function which waits for the
notification from the `didSignIn` delegate method. Instead we can just call the
auth callback directly.

Also remove the entire `registerCallback` `didSignIn` code to avoid future confusion.

Testing done:
- Without this, `getJWT` did not return anything since we had removed the
`registerCallback` call in
https://github.com/e-mission/cordova-jwt-auth/pull/45/commits/458500664df028c898867f409dca27ebe9c87c6d

- With this, it did