:fire: Remove the old UserProfile and the related migration path
In 874ac6ea1d9a54663dbd977ac446fb54d1ec707b ("Modify prompted auth storage part
1 (get + migrate)"), we switched from storing the token in shared storage
to the database. This was prompted by reports that the token was not stored
correctly on multiple iOS and android phones.

For backwards compatibility, we also added a migration path that would copy the
token from the user profile into the database.

The backwards compat path is now obsolete - the CanBikeCO app has been removed
from the stores, and NREL OpenPATH has always stored the information in the
database.

So let's remove the backward compat code path and the related `UserProfile` class.

Testing done:
- Builds on both android and iOS