:bug: change `PromptedAuth` to actually use the new interface
I am not sure why this was not caught as a compile time check.

As part of 0ef812ce7a5252785621657b61a22aff078abb30, we changed the interface
for the auth creator to remove a lot of the callbacks and to have a simple
`getOPCode/setOPCode` interface.

However, we didn't change the implementation in PromptedAuth to match.
This should have been caught at compile time, since PromptedAuth implements the
AuthTokenCreator protocol, but it wasn't.

This led to an error while reading the token
https://github.com/e-mission/e-mission-phone/pull/927#issuecomment-1458465391

This finally changes PromptedAuth to match the new interface

Testing done:
- Builds
- We can now see the token and also re-login with it