:fire: Remove all library dependencies on android and iOS
Since we no longer support google and openid auth, we can remove the libraries
that support them. Also remove a couple of other frameworks which don't seem to
make a difference.

Note that removing `<pod name="GoogleSignIn" spec="~> 6.0.2" />` breaks the
build because then the comm helper on iOS does not find the
`GTMSessionFetcher`, which is otherwise imported as a dependency of the `GoogleSignIn`

We add it as a dependency to the server communication plugin instead, which works